refactor: :art: Created folders for better organization
