+++ src/automaton/NFA.py
